requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-filter==2.1.0
django-multiselectfield==0.1.8
idna==2.8
Pillow==6.0.0
pytz==2019.1
requests==2.21.0
sqlparse==0.3.0
stripe==2.27.0
urllib3==1.24.3
